Fixed apprequests dialog support on Android
The previous implementation of DialogActivity assumed a feed dialog and processed the response accordingly, but Facebook supports 3 dialog types ("feed", "apprequests", and "oauth"). The iOS version does not have this limitation.

This update tracks the type of dialog invoked (both "feed" and "apprequests" are reported), and process the response to keep the return data consistent with the iOS version.

This fix is backwards compatible.